+++ lib/vect3.h
@@ -3,9 +3,9 @@
 #include <math.h>
 
 typedef struct {
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 } Vect3;
 
 typedef Vect3 Point;
@@ -17,15 +17,15 @@ Vect3 Vect3Add(Vect3 v1, Vect3 v2);
 
 Vect3 Vect3Neg(Vect3 v);
 
-void Vect3ScaleP(Vect3* v1, double s);
-Vect3 Vect3Scale(Vect3 v1, double s);
+void Vect3ScaleP(Vect3* v1, float s);
+Vect3 Vect3Scale(Vect3 v1, float s);
 
-double Vect3Len(Vect3 v1);
+float Vect3Len(Vect3 v1);
 
 void Vect3NormP(Vect3* v1);
 Vect3 Vect3Norm(Vect3 v1);
 
-double Vect3Dot(Vect3 v1, Vect3 v2);
+float Vect3Dot(Vect3 v1, Vect3 v2);
 Vect3 Vect3Cross(Vect3 v1, Vect3 v2);
 
 #endif
